int -> size_t where appropriate

+++ os/include/os/util.h
@@ -80,10 +80,10 @@ size_t str_writer(void *dest, const char *data);
 #endif
 
 int nitoa(int i, int base, char *dest, size_t len);
-int pandos_snprintf(char *dest, size_t len, const char *fmt, ...);
+size_t pandos_snprintf(char *dest, size_t len, const char *fmt, ...);
 
 #ifndef __x86_64__
-int pandos_fprintf(int fd, const char *fmt, ...);
+size_t pandos_fprintf(int fd, const char *fmt, ...);
 #define pandos_printf(...) pandos_fprintf(0, __VA_ARGS__)
 #endif
